Feat: Solve quiz 4
Following are the question which

a. What is the probability of an individual converting regardless of the page they receive?

b. Given that an individual was in the control group, what is the probability they converted?

c. Given that an individual was in the treatment group, what is the probability they converted?

d. What is the probability that an individual received the new page?